rename index.js to PublishNotebookPane
